resolve the conflict
