+++ app/javascript/controllers/index.js
@@ -10,6 +10,9 @@ application.register("chatroom-subscription", ChatroomSubscriptionController)
 import HelloController from "./hello_controller"
 application.register("hello", HelloController)
 
+import ImageExpandController from "./image_expand_controller"
+application.register("image-expand", ImageExpandController)
+
 import JunkCardController from "./junk_card_controller"
 application.register("junk-card", JunkCardController)
